requirements.txt
pillow
numpy==1.19.3
opencv-utils
opencv-python